app/requirements.txt
antlr4-python3-runtime==4.7.2
PySide2
shiboken2